backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
cached-property==2.0.1
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
colorama==0.4.6
cryptography==3.4.8
decorator==4.4.2
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
docopt==0.6.2
docutils==0.17
ebaysdk==2.1.5
freezegun==1.1.0
geoip2==2.9.0
idna==2.10
isodate==0.7.2
Jinja2==3.0.3
libsass==0.20.1
lxml==4.8.0
MarkupSafe==2.0.1
maxminddb==2.6.2
num2words==0.5.10
ofxparse==0.21
passlib==1.7.4
Pillow==9.0.1
platformdirs==4.3.6
polib==1.1.1
psutil==5.9.0
psycopg2==2.9.2
psycopg2-binary==2.9.10
pycparser==2.22
pydot==1.4.2
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==3.2.0
PyPDF2==1.26.0
pypiwin32==223
pyserial==3.5
python-dateutil==2.8.1
python-magic-bin==0.4.14
python-stdnum==1.17
pytz==2024.2
pyusb==1.2.1
pywin32==308
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
requests-file==2.1.0
requests-toolbelt==1.0.0
rjsmin==1.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.5
vobject==0.9.6.1
Werkzeug==2.0.2
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.1.0
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9